refactor(recent-food): move module to diet subdirectory
