+++ src/modules/diet/recent-food/application/usecases/touchRecentFood.ts
@@ -1,10 +1,10 @@
 import { authUseCases } from '~/modules/auth/application/usecases/authUseCases'
-import { recentFoodCrudService } from '~/modules/recent-food/application/usecases/deps'
-import type { RecentFoodReference } from '~/modules/recent-food/application/usecases/extractRecentFoodReference'
+import { recentFoodCrudService } from '~/modules/diet/recent-food/application/usecases/deps'
+import type { RecentFoodReference } from '~/modules/diet/recent-food/application/usecases/extractRecentFoodReference'
 import {
   createNewRecentFood,
   type NewRecentFood,
-} from '~/modules/recent-food/domain/recentFood'
+} from '~/modules/diet/recent-food/domain/recentFood'
 
 export async function touchRecentFood(recentFoodRef: RecentFoodReference) {
   const currentRecentFood =
