+++ src/modules/diet/recent-food/domain/recentFoodRepository.ts
@@ -1,8 +1,8 @@
-import type { Template } from '~/modules/diet/template/domain/template'
 import {
   type NewRecentFood,
   type RecentFood,
-} from '~/modules/recent-food/domain/recentFood'
+} from '~/modules/diet/recent-food/domain/recentFood'
+import type { Template } from '~/modules/diet/template/domain/template'
 import { type User } from '~/modules/user/domain/user'
 
 export type RecentFoodRepository = {
