+++ src/modules/diet/recent-food/infrastructure/recentFoodRepository.ts
@@ -1,10 +1,10 @@
-import type { Template } from '~/modules/diet/template/domain/template'
 import {
   type NewRecentFood,
   type RecentFood,
-} from '~/modules/recent-food/domain/recentFood'
-import { type RecentFoodRepository } from '~/modules/recent-food/domain/recentFoodRepository'
-import { type RecentFoodGateway } from '~/modules/recent-food/infrastructure/supabase/supabaseRecentFoodGateway'
+} from '~/modules/diet/recent-food/domain/recentFood'
+import { type RecentFoodRepository } from '~/modules/diet/recent-food/domain/recentFoodRepository'
+import { type RecentFoodGateway } from '~/modules/diet/recent-food/infrastructure/supabase/supabaseRecentFoodGateway'
+import type { Template } from '~/modules/diet/template/domain/template'
 import { type User } from '~/modules/user/domain/user'
 import { logging } from '~/shared/utils/logging'
