+++ src/modules/diet/recent-food/infrastructure/supabase/supabaseRecentFoodGateway.ts
@@ -3,14 +3,14 @@ import { z } from 'zod/v4'
 import { foodSchema } from '~/modules/diet/food/domain/food'
 import { supabaseItemMapper } from '~/modules/diet/item/infrastructure/supabase/supabaseItemMapper'
 import { supabaseMacroNutrientsMapper } from '~/modules/diet/macro-nutrients/infrastructure/supabase/supabaseMacroNutrientsMapper'
-import { recipeSchema } from '~/modules/diet/recipe/domain/recipe'
-import { type Template } from '~/modules/diet/template/domain/template'
 import {
   type NewRecentFood,
   type RecentFood,
-} from '~/modules/recent-food/domain/recentFood'
-import { SUPABASE_TABLE_RECENT_FOODS } from '~/modules/recent-food/infrastructure/supabase/constants'
-import { supabaseRecentFoodMapper } from '~/modules/recent-food/infrastructure/supabase/supabaseRecentFoodMapper'
+} from '~/modules/diet/recent-food/domain/recentFood'
+import { SUPABASE_TABLE_RECENT_FOODS } from '~/modules/diet/recent-food/infrastructure/supabase/constants'
+import { supabaseRecentFoodMapper } from '~/modules/diet/recent-food/infrastructure/supabase/supabaseRecentFoodMapper'
+import { recipeSchema } from '~/modules/diet/recipe/domain/recipe'
+import { type Template } from '~/modules/diet/template/domain/template'
 import { type User } from '~/modules/user/domain/user'
 import { type Json } from '~/shared/supabase/database.types'
 import { supabase } from '~/shared/supabase/supabase'
